Package structure: moved aggregated statistics.
Part of #963.
