+++ src/main/AndroidManifest.xml
@@ -316,7 +316,7 @@ limitations under the License.
 
         <activity android:name=".TrackEditActivity" />
 
-        <activity android:name=".AggregatedStatisticsActivity" />
+        <activity android:name=".ui.aggregatedStatistics.AggregatedStatisticsActivity" />
 
         <activity android:name=".io.file.importer.DirectoryChooserActivity$ExportDirectoryChooserActivity" />
         <activity android:name=".io.file.importer.DirectoryChooserActivity$DefaultTrackExportDirectoryChooserActivity" />
