+++ src/main/java/de/dennisguse/opentracks/AbstractTrackDeleteActivity.java
@@ -23,10 +23,10 @@ import java.util.Arrays;
 import java.util.stream.Collectors;
 
 import de.dennisguse.opentracks.content.data.Track;
-import de.dennisguse.opentracks.fragments.ConfirmDeleteDialogFragment;
-import de.dennisguse.opentracks.fragments.ConfirmDeleteDialogFragment.ConfirmDeleteCaller;
 import de.dennisguse.opentracks.services.TrackDeleteService;
 import de.dennisguse.opentracks.services.TrackDeleteServiceConnection;
+import de.dennisguse.opentracks.ui.aggregatedStatistics.ConfirmDeleteDialogFragment;
+import de.dennisguse.opentracks.ui.aggregatedStatistics.ConfirmDeleteDialogFragment.ConfirmDeleteCaller;
 
 /**
  * An abstract class for the following common tasks across
