+++ src/main/java/de/dennisguse/opentracks/ui/aggregatedStatistics/FilterDialogFragment.java
@@ -1,4 +1,4 @@
-package de.dennisguse.opentracks.fragments;
+package de.dennisguse.opentracks.ui.aggregatedStatistics;
 
 import android.app.Dialog;
 import android.content.Context;
